RestAdapter and service instances should be singletons.
