+++ tv/src/main/java/com/jerrellmardis/amphitheatre/api/TMDbClient.java
@@ -16,23 +16,13 @@
 
 package com.jerrellmardis.amphitheatre.api;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.jerrellmardis.amphitheatre.BuildConfig;
 import com.jerrellmardis.amphitheatre.model.tmdb.Config;
 import com.jerrellmardis.amphitheatre.model.tmdb.Metadata;
 import com.jerrellmardis.amphitheatre.model.tmdb.Movie;
-import com.jerrellmardis.amphitheatre.util.ApiConstants;
 
-import retrofit.RequestInterceptor;
-import retrofit.RestAdapter;
-import retrofit.converter.GsonConverter;
 import retrofit.http.GET;
 import retrofit.http.Path;
 import retrofit.http.Query;
-
-import static retrofit.RestAdapter.LogLevel.FULL;
-import static retrofit.RestAdapter.LogLevel.NONE;
 
 /**
  * Created by Jerrell Mardis on 7/8/14.
@@ -50,34 +40,25 @@ public class TMDbClient {
         Movie getMovie(@Path("id") Long id);
     }
 
+    private static TMDbService service;
+
+    private static TMDbService getService() {
+        if (service == null) {
+            service = Clients.getRestAdapter().create(TMDbService.class);
+        }
+        return service;
+    }
+
     public static Config getConfig() {
-        return getHttpService().getConfig();
+        return getService().getConfig();
     }
 
     public static Metadata getMetadata(CharSequence name, CharSequence year) {
-        return getHttpService().getMetadata(name, year);
+        return getService().getMetadata(name, year);
     }
 
     public static Movie getMovie(Long id) {
-        return getHttpService().getMovie(id);
+        return getService().getMovie(id);
     }
 
-    private static TMDbService getHttpService() {
-        Gson gson = new GsonBuilder().create();
-
-        RestAdapter restAdapter = new RestAdapter.Builder()
-                .setConverter(new GsonConverter(gson))
-                .setEndpoint(ApiConstants.TMDB_SERVER_URL)
-                .setRequestInterceptor(new RequestInterceptor() {
-                    @Override
-                    public void intercept(RequestFacade request) {
-                        request.addQueryParam("api_key", ApiConstants.TMDB_SERVER_API_KEY);
-                    }
-                })
-                .build();
-
-        restAdapter.setLogLevel(BuildConfig.DEBUG ? FULL : NONE);
-
-        return restAdapter.create(TMDbService.class);
-    }
 }
